refactor(exceptions): make catching exceptions more generalized and proper
