+++ src/shared/filters/http-exception.filter.ts
@@ -1,0 +1,27 @@
+import { ExceptionFilter, Catch, HttpException, ArgumentsHost, HttpStatus } from '@nestjs/common';
+
+@Catch()
+export class HttpExceptionFilter implements ExceptionFilter {
+  catch(error: Error, host: ArgumentsHost) {
+    const ctx = host.switchToHttp();
+    const request = ctx.getRequest();
+    const response = ctx.getResponse();
+    const status = (error instanceof HttpException) ? error.getStatus() : HttpStatus.INTERNAL_SERVER_ERROR;
+
+    if (status === HttpStatus.INTERNAL_SERVER_ERROR) {
+      response.status(500).json({
+        statusCode: 500,
+        timestamp: new Date().toISOString(),
+        path: request.url,
+        message: String(error.stack)
+      })
+    }
+
+    response.status(status).json({
+      statusCode: status,
+      timestamp: new Date().toISOString(),
+      path: request.url,
+      message: error.message
+    })
+  }
+}
