+++ src/shared/filters/index.ts
@@ -1,2 +1,2 @@
 export * from './validation-exception.filter'
-export * from './internal-server-exception.filter'
+export * from './http-exception.filter'
