--- src/shared/filters/internal-server-exception.filter.ts
@@ -1,16 +1,0 @@
-import { ExceptionFilter, Catch, HttpException, ArgumentsHost, HttpStatus } from '@nestjs/common';
-
-@Catch()
-export class InternalServerExceptionFilter implements ExceptionFilter {
-  catch(error: Error, host: ArgumentsHost) {
-    let response = host.switchToHttp().getResponse();
-    let status = (error instanceof HttpException) ? error.getStatus() : HttpStatus.INTERNAL_SERVER_ERROR;
-
-    if (status === HttpStatus.INTERNAL_SERVER_ERROR) {
-      response.status(500).json({
-        statusCode: 500,
-        message: String(error.stack)
-      })
-    }
-  }
-}
